+++ frontends/election-manager/src/components/print_button.test.tsx
@@ -1,128 +1,80 @@
 import React from 'react';
-import {
-  act,
-  fireEvent,
-  waitFor,
-  getByText as domGetByText,
-  waitForElementToBeRemoved,
-} from '@testing-library/react';
-import {
-  fakeKiosk,
-  fakePrinter,
-  fakePrinterInfo,
-} from '@votingworks/test-utils';
-import { deferred } from '@votingworks/utils';
+import { waitFor, screen, within, render } from '@testing-library/react';
+import { advancePromises, fakeKiosk } from '@votingworks/test-utils';
 
+import userEvent from '@testing-library/user-event';
 import { PrintButton } from './print_button';
 import { renderInAppContext } from '../../test/render_in_app_context';
 
-beforeAll(() => {
-  window.kiosk = fakeKiosk();
+beforeEach(() => {
+  jest.useFakeTimers();
 });
 
-afterAll(() => {
+test('happy path flow', async () => {
+  window.kiosk = fakeKiosk();
+  const mockPrint = jest.fn();
+  renderInAppContext(<PrintButton print={mockPrint}>Print</PrintButton>, {
+    hasPrinterAttached: true,
+  });
+  expect(screen.queryByRole('alertdialog')).not.toBeInTheDocument();
+  userEvent.click(screen.getByRole('button', { name: 'Print' }));
+  within(screen.getByRole('alertdialog')).getByText('Printing');
+  expect(mockPrint).toHaveBeenCalledTimes(1);
+  jest.advanceTimersByTime(3000);
+  await waitFor(() => {
+    expect(screen.queryByText('Printing')).not.toBeInTheDocument();
+  });
   delete window.kiosk;
 });
 
-test('if only disconnected printers, show error modal', async () => {
-  const mockKiosk = window.kiosk! as jest.Mocked<KioskBrowser.Kiosk>;
-  mockKiosk.getPrinterInfo.mockResolvedValue([
-    fakePrinterInfo({ connected: false }),
-    fakePrinterInfo({ name: 'VxPrinter', connected: false }),
-  ]);
-
-  const printer = fakePrinter();
-  const afterPrint = jest.fn();
-  const { getByText } = renderInAppContext(
-    <PrintButton sides="two-sided-long-edge" afterPrint={afterPrint}>
-      Print Now
-    </PrintButton>,
-    { printer }
-  );
-
-  await act(async () => {
-    fireEvent.click(getByText('Print Now'));
-
-    await waitFor(() =>
-      getByText('The printer is not connected', { exact: false })
-    );
+test('does not expect a printer attached in browser mode', async () => {
+  const mockPrint = jest.fn();
+  renderInAppContext(<PrintButton print={mockPrint}>Print</PrintButton>, {
+    hasPrinterAttached: false,
   });
-
-  expect(mockKiosk.getPrinterInfo).toBeCalled();
-
-  expect(printer.print).not.toBeCalled();
-  expect(afterPrint).not.toBeCalled();
-});
-
-test('if connected printers, show printing modal', async () => {
-  const mockKiosk = window.kiosk! as jest.Mocked<KioskBrowser.Kiosk>;
-  mockKiosk.getPrinterInfo.mockResolvedValue([
-    fakePrinterInfo({ connected: false }),
-    fakePrinterInfo({ name: 'VxPrinter', connected: true }),
-  ]);
-
-  const printer = fakePrinter();
-  const afterPrint = jest.fn();
-  const { getByText } = renderInAppContext(
-    <PrintButton afterPrint={afterPrint} sides="two-sided-long-edge">
-      Print Now
-    </PrintButton>,
-    { printer }
-  );
-
-  await act(async () => {
-    fireEvent.click(getByText('Print Now'));
-
-    await waitFor(() => getByText('Printing', { exact: false }));
-  });
-
-  expect(mockKiosk.getPrinterInfo).toBeCalled();
-
-  expect(printer.print).toBeCalled();
-  expect(afterPrint).toBeCalled();
-});
-
-test('if passed a printTarget, render for printing', async () => {
-  const mockKiosk = window.kiosk! as jest.Mocked<KioskBrowser.Kiosk>;
-  mockKiosk.getPrinterInfo.mockResolvedValue([
-    fakePrinterInfo({ connected: false }),
-    fakePrinterInfo({ name: 'VxPrinter', connected: true }),
-  ]);
-
-  const printer = fakePrinter();
-  const printTarget = <p>Print Target</p>;
-  const printTargetTestId = 'print-target';
-  const { getByTestId, getByText, queryByText } = renderInAppContext(
-    <PrintButton
-      sides="two-sided-long-edge"
-      printTarget={printTarget}
-      printTargetTestId={printTargetTestId}
-    >
-      Print Now
-    </PrintButton>,
-    { printer }
-  );
-
-  const printDeferred = deferred<void>();
-  printer.print.mockReturnValueOnce(printDeferred.promise);
-  fireEvent.click(getByText('Print Now'));
-
-  // Check that printTarget is rendered before printing
+  userEvent.click(screen.getByRole('button', { name: 'Print' }));
   await waitFor(() => {
-    const renderedPrintTarget = getByTestId('print-target');
-    domGetByText(renderedPrintTarget, 'Print Target');
-    expect(printer.print).not.toBeCalled();
+    expect(mockPrint).toHaveBeenCalled();
   });
-
-  // Check that printTarget is still rendered after printing
-  await waitFor(() => {
-    expect(printer.print).toBeCalled();
-    getByText('Print Target');
-  });
-
-  // Allow `printer.print` to resolve
-  printDeferred.resolve();
-
-  // Check that printTarget is removed after printing
-  await waitForElementToBeRemoved(() => queryByText('Print Target'));
 });
+
+test('has option to not show the default progress modal', async () => {
+  const mockPrint = jest.fn();
+  render(
+    <PrintButton print={mockPrint} useDefaultProgressModal={false}>
+      Print
+    </PrintButton>
+  );
+  userEvent.click(screen.getByRole('button', { name: 'Print' }));
+  expect(mockPrint).toHaveBeenCalled();
+  expect(screen.queryByRole('alertdialog')).not.toBeInTheDocument();
+  expect(screen.queryByText('Printing')).not.toBeInTheDocument();
+
+  // to prevent test warnings for state updates after teardown
+  await advancePromises();
+});
+
+test('shows printer not connected warning', () => {
+  window.kiosk = fakeKiosk();
+  const mockPrint = jest.fn();
+  renderInAppContext(
+    <PrintButton print={mockPrint} useDefaultProgressModal={false}>
+      Print
+    </PrintButton>,
+    { hasPrinterAttached: false }
+  );
+  userEvent.click(screen.getByRole('button', { name: 'Print' }));
+  const modal = screen.getByRole('alertdialog');
+  within(modal).getByText('The printer is not connected.');
+  expect(mockPrint).not.toHaveBeenCalled();
+  expect(
+    within(modal).queryByRole('button', { name: 'Continue' })
+  ).toBeDisabled();
+  userEvent.click(within(modal).getByRole('button', { name: 'Close' }));
+  expect(screen.queryByRole('alertdialog')).not.toBeInTheDocument();
+
+  delete window.kiosk;
+});
+
+// The path in which you connect a printer and select "Continue" from the
+// PrinterNotConnectedModal is covered in app.test.tsx
